Refractor member session system
Since the member session system was improperly developed with the data being saved on the disk, I changed taht to save it inside of a static field. Member data can now easily be gottenw ith Member.getInstance() and the data gathered by that will always be changing as the database updates.